Split algorithms into multiple files
